Allowed multiple unit sizes, and created the sample size variable
Changed the default size for all stars to be 0.01 and units back in parsec so we can see the stars in the star cluster in the same FOV. Allowed the unit size (in pc) to be a variable, and the step for how many star data points are gathered to be a variable. This allows modifying the play behaviour without modifying the scripts.